requirements.txt
torch==2.2.2
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
timm==1.0.13
pytorch-lightning==2.5.0.post0
hydra-core==1.3.2
wandb==0.19.2
invoke==2.2.0
opendatasets==0.1.22
pandas==2.2.3
numpy==1.26.4
coverage==7.6.10
pytest==8.3.4
kagglehub==0.3.6
ruff==0.9.0
gdown==5.2.0
tqdm==4.67.1
loguru==0.7.3
onnx==1.17.0
onnxruntime==1.20.1
onnxscript==0.1.0.dev20250114
bentoml==1.3.19
streamlit==1.41.1
google_cloud_run==0.10.14
